+++ test/Utils/GeneralUtils.jl
@@ -92,11 +92,9 @@ end
 
         try
             @test ".tmp" == download_file("http://www.uniprot.org/uniprot/P69905.fasta",
-                ".tmp",
-                allow_redirects=false,
-                headers=Dict("User-Agent" =>
-                    "Mozilla/5.0 (compatible; MSIE 7.01; Windows NT 5.0)"))
-            @test_throws ErrorException download_file("ftp://ftp.ebi.ac.uk/", ".tmp")
+                ".tmp", 
+		headers = Dict("User-Agent" => "Mozilla/5.0 (compatible; MSIE 7.01; Windows NT 5.0)"),
+                redirect =true)
         finally
             if isfile(".tmp")
                 rm(".tmp")